deleted edit-room.vue and category.vue, the authentication is missing
