--- src/components/category-bar.vue
@@ -1,14 +1,0 @@
-<script setup>
-
-</script>
-<template>
-  <h1>test</h1>
-</template>
-
-<style scoped>
-*{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-</style>
