--- src/components/categoryBar.vue
@@ -1,15 +1,0 @@
-<script setup>
-
-
-</script>
-<template>
-    <h4>hi</h4>
-</template>
-
-<style scoped>
-    *{
-        margin: 0;
-        padding: 0;
-        box-sizing: border-box;
-    }
-</style>
